+++ CHANGELOG.md
@@ -4,6 +4,18 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased] Unreleased
 
+## [2.3.2] 2020-03-29;
+
+### Fixed
+
+- absolute paths handling in the `configFile` parameter of `WPLoader` configuration
+
+## [2.3.1] 2020-03-29;
+
+### Fixed
+
+- absolute paths handling in the `pluginsFolder` parameter of `WPLoader` configuration
+
 ## [2.3.0] 2020-03-29;
 
 ### Added
@@ -1234,4 +1246,6 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 [2.2.36]: https://github.com/lucatume/wp-browser/compare/2.2.35...2.2.36
 [2.2.37]: https://github.com/lucatume/wp-browser/compare/2.2.36...2.2.37
 [2.3.0]: https://github.com/lucatume/wp-browser/compare/2.2.37...2.3.0
-[unreleased]: https://github.com/lucatume/wp-browser/compare/2.3.0...HEAD
+[2.3.1]: https://github.com/lucatume/wp-browser/compare/2.3.0...2.3.1
+[2.3.2]: https://github.com/lucatume/wp-browser/compare/2.3.1...2.3.2
+[unreleased]: https://github.com/lucatume/wp-browser/compare/2.3.2...HEAD